+++ data-raw/clean_raw_data.R
@@ -1,17 +1,34 @@
-# surnames_race <- get_surnames_race()
-# usethis::use_data(surnames_race, internal = TRUE, overwrite = TRUE)
-# usethis::use_data(surnames_race, internal = FALSE, overwrite = TRUE)
-
-# get_surnames_race <- function(){
-#   # https://www.census.gov/topics/population/genealogy/data/2010_surnames.html
-#   data <- read.csv("data-raw/Names_2010Census.csv")
-#   data <- clean_data(data)
-#   data_2000 <- readxl::read_excel("data-raw/app_c.xlsx", sheet = 1)
-#   data_2000 <- clean_data(data_2000)
+# # surnames_race <- get_surnames_race()
+#
+# # https://www.census.gov/topics/population/genealogy/data/2010_surnames.html
+# data <- read.csv("data-raw/Names_2010Census.csv")
+# data <- clean_data(data)
+# data_2000 <- readxl::read_excel("data-raw/app_c.xlsx", sheet = 1)
+# data_2000 <- clean_data(data_2000)
+# data <-
+#   data %>%
+#   dplyr::bind_rows(data_2000)
+# surnames_race <- get_surnames_race(data)
+#
+# first_names <- readxl::read_excel("data-raw/firstnames.xlsx", sheet = 2)
+# names(first_names) <- gsub("^obs$", "count", names(first_names))
+# names(first_names) <- gsub("^firstname$", "name", names(first_names))
+# first_names <- clean_data(first_names)
+# first_names_race <- get_surnames_race(first_names)
+#
+#
+#
+# first_names_gender <- get_ssn_gender()
+#
+# usethis::use_data(first_names_race, surnames_race, first_names_gender,
+#                   internal = TRUE, overwrite = TRUE)
+# usethis::use_data(first_names_race, surnames_race, first_names_gender,
+#                   internal = FALSE, overwrite = TRUE)
+# #
+# get_surnames_race <- function(data){
 #
 #   data <-
 #     data %>%
-#     dplyr::bind_rows(data_2000) %>%
 #     dplyr::group_by(name) %>%
 #     dplyr::summarize_all(sum) %>%
 #     dplyr::mutate(probability_white           = count_white / count,
@@ -29,7 +46,7 @@
 #                   -count_2races,
 #                   -count_hispanic,
 #                   -count
-#                   )
+#     )
 #
 #   data$likely_race <- ""
 #   race_columns <-
@@ -40,15 +57,15 @@
 #       "probability_2races",
 #       "probability_hispanic")
 #   race_columns_permutations <- combinat::permn(race_columns)
+#   message(length(race_columns_permutations))
 #   for (i in 1:length(race_columns_permutations)) {
 #
 #     race_columns_temp <- race_columns_permutations[[i]]
-#
 #     data$likely_race <- paste0(data$likely_race, ", ",
 #                                race_columns_temp[max.col(data[, race_columns_temp],
 #                                                          ties.method = "first")])
 #
-#     z<- strsplit(data$likely_race, ", ")
+#     z <- strsplit(data$likely_race, ", ")
 #     head(z)
 #     z <- sapply(z, unique, simplify = FALSE)
 #     z <- sapply(z, sort, simplify = FALSE)
@@ -74,7 +91,7 @@
 #                         probability_hispanic,
 #                         probability_white,
 #                         probability_2races) %>%
-# dplyr::mutate_if(is.numeric, round, 4)
+#     dplyr::mutate_if(is.numeric, round, 4)
 #   data <- as.data.frame(data)
 #
 #   return(data)
@@ -128,3 +145,42 @@
 #   return(data)
 # }
 #
+
+#
+# get_ssn_gender <- function(){
+#   # Get social security name/gender data
+#   files <- list.files("data-raw", pattern = "txt")
+#   data  <- data.frame()
+#   for (i in 1:length(files)) {
+#     temp <- read.table(paste0("data-raw/", files[i]), header = FALSE,
+#                        sep = ",")
+#     data <- dplyr::bind_rows(data, temp)
+#   }
+#   names(data) <- c("name", "gender", "count")
+#   data$gender <- gsub("^M$", "male", data$gender)
+#   data$gender <- gsub("^F$", "female", data$gender)
+#
+#   data <-
+#     data %>%
+#     dplyr::group_by(name,
+#                     gender) %>%
+#     dplyr::summarize(count = sum(count)) %>%
+#     dplyr::arrange(name) %>%
+#     dplyr::ungroup() %>%
+#     dplyr::mutate(name = tolower(name))
+#
+#
+#   data <- data.frame(reshape::cast(data,
+#                                   name ~ gender, value = c("count"), fun = sum)) %>%
+#     dplyr::mutate(total              = female + male,
+#                   probability_male   = male / total,
+#                   probability_female = female / total) %>%
+#     dplyr::select(-female,
+#                   -male,
+#                   -total)
+#
+#   data$likely_gender <- "female"
+#   data$likely_gender[data$probability_male >  data$probability_female] <- "male"
+#   data$likely_gender[data$probability_male == data$probability_female] <- "female, male"
+#   return(data)
+# }