+++ tests/testthat/test-race.R
@@ -6,6 +6,14 @@ rlincoln      <- predictrace::predict_race("RLincoln")
 perez         <- predictrace::predict_race("Perez")
 lucky         <- predictrace::predict_race("lucky")
 
+
+abbey         <- predictrace::predict_race("abbey", surname = FALSE)
+anibal_kenny  <- predictrace::predict_race(c("anibal", "KennY"), surname = FALSE)
+harry_houdini <- predictrace::predict_race("harry houdini", surname = FALSE)
+zzzdoan       <- predictrace::predict_race("ZZZDoan", surname = FALSE)
+doan          <- predictrace::predict_race("Doan", surname = FALSE)
+cedric        <- predictrace::predict_race("cedric", surname = FALSE)
+
 test_that("predict_race works", {
 
   expect_named(smith, c("name",
@@ -56,7 +64,6 @@ test_that("predict_race works", {
   expect_equal(smith_lincoln$probability_2races,
                c(0.0191, 0.025))
 
-
   expect_true(is.na(ben_smith$likely_race))
   expect_true(is.na(predict_race("Ben smith", probability = FALSE)$likely_race))
   expect_true(is.na(ben_smith$probability_american_indian))
@@ -65,8 +72,6 @@ test_that("predict_race works", {
   expect_true(is.na(ben_smith$probability_hispanic))
   expect_true(is.na(ben_smith$probability_white))
   expect_true(is.na(ben_smith$probability_2races))
-
-
 
   expect_equal(lincoln$likely_race, "white")
   expect_equal(predict_race("Lincoln", probability = FALSE)$likely_race, "white")
@@ -85,5 +90,85 @@ test_that("predict_race works", {
   expect_true(is.na(rlincoln$probability_hispanic))
   expect_true(is.na(rlincoln$probability_white))
   expect_true(is.na(rlincoln$probability_2races))
+
+})
+
+
+test_that("predict_race for first name works", {
+
+
+  expect_named(abbey, c("name",
+                        "match_name",
+                        "likely_race",
+                        "probability_american_indian",
+                        "probability_asian",
+                        "probability_black",
+                        "probability_hispanic",
+                        "probability_white",
+                        "probability_2races"))
+  expect_named(predict_race("abbey", probability = FALSE, surname = FALSE), c("name",
+                                                             "match_name",
+                                                             "likely_race"))
+  expect_equal(abbey$likely_race, "white")
+
+  expect_equal(doan$likely_race, "asian")
+  expect_equal(predict_race("doan", probability = FALSE)$likely_race, "asian")
+  expect_equal(doan$probability_american_indian, 0)
+  expect_equal(doan$probability_asian, 0.9667)
+  expect_equal(doan$probability_black, 0)
+  expect_equal(doan$probability_hispanic, 0)
+  expect_equal(doan$probability_white, 0.0333)
+  expect_equal(doan$probability_2races, 0)
+
+
+  expect_equal(abbey$probability_american_indian, 0)
+  expect_equal(abbey$probability_asian, 0)
+  expect_equal(abbey$probability_black, 0.0351)
+  expect_equal(abbey$probability_hispanic, 0)
+  expect_equal(abbey$probability_white, 0.9649)
+  expect_equal(abbey$probability_2races, 0)
+
+  expect_equal(anibal_kenny$likely_race, c("hispanic", "white"))
+  expect_equal(predict_race(c("anibal", "KennY"), probability = FALSE, surname = FALSE)$likely_race,
+               c("hispanic", "white"))
+  expect_equal(anibal_kenny$probability_american_indian,
+               c(0, 0.0029))
+  expect_equal(anibal_kenny$probability_asian,
+               c(0, .3736))
+  expect_equal(anibal_kenny$probability_black,
+               c(0, 0.0402))
+  expect_equal(anibal_kenny$probability_hispanic,
+               c(.8889, 0.0287))
+  expect_equal(anibal_kenny$probability_white,
+               c(.1111, 0.5488))
+  expect_equal(anibal_kenny$probability_2races,
+               c(0, 0.0058))
+
+  expect_true(is.na(harry_houdini$likely_race))
+  expect_true(is.na(predict_race("harry houdini", probability = FALSE, surname = FALSE)$likely_race))
+  expect_true(is.na(harry_houdini$probability_american_indian))
+  expect_true(is.na(harry_houdini$probability_asian))
+  expect_true(is.na(harry_houdini$probability_black))
+  expect_true(is.na(harry_houdini$probability_hispanic))
+  expect_true(is.na(harry_houdini$probability_white))
+  expect_true(is.na(harry_houdini$probability_2races))
+
+  expect_equal(cedric$likely_race, "black")
+  expect_equal(predict_race("cedric", probability = FALSE, surname = FALSE)$likely_race, "black")
+  expect_equal(cedric$probability_american_indian, 0)
+  expect_equal(cedric$probability_asian, 0.0909)
+  expect_equal(cedric$probability_black, 0.6212)
+  expect_equal(cedric$probability_hispanic, 0)
+  expect_equal(cedric$probability_white, 0.2727)
+  expect_equal(cedric$probability_2races, 0.0151)
+
+  expect_true(is.na(zzzdoan$likely_race))
+  expect_true(is.na(predict_race("zzzdoan", probability = FALSE, surname = FALSE)$likely_race))
+  expect_true(is.na(zzzdoan$probability_american_indian))
+  expect_true(is.na(zzzdoan$probability_asian))
+  expect_true(is.na(zzzdoan$probability_black))
+  expect_true(is.na(zzzdoan$probability_hispanic))
+  expect_true(is.na(zzzdoan$probability_white))
+  expect_true(is.na(zzzdoan$probability_2races))
 
 })