+++ NEWS.md
@@ -1,3 +1,11 @@
+# predictrace 2.0.0
+
+* Adds parameter in `predict_race()` where you can enter a first name and get the 
+  likely race of that first name.
+* Adds function `predict_gender()` that uses United States Social Security 
+  Administration (SSA) data from 1880-2019 to predict the gender of a first
+  name.
+
 # predictrace 1.2.1
 
 * Combines 2010 Census data and 2000 Census data together.